refactor(frontend): split playtest model rendering
